Update cometbft-proto to protobufs in CometBFT v1.0 (pre-release) (#7)
* proto-compiler: use single CometBFT checkout

Generate from the new structure with versioned proto files.

* proto: remove tendermint modules

These are no longer going to be generated with the switch to versioned
cometbft.* protos.

* proto: re-generate cometbft.*, remove tendermint.*

Bulk change committing output of the recently changed proto-compiler.

* proto: update serde support for Evidence, PublicKey

The generated struct types only come from cometbft::* modules
generated from versioned protobufs.
Update the imports and manually written implementations.

* cometbft: update proto conversions

Convert imports and the provided conversions to/from the protobuf
structs to the version structure of cometbft_proto.

* cometbft: restructure versioned Request and Response in abci

Move the versions under the abci module and make the names match
the proto package names.

* proto: update imports in tests

* rpc: rename and update dialects

The dialect of 0.37 is now known as 1.0.

* proto: generate serde for abci::v1::InfoResponse

Add a flag for proto-compiler that was not copied from abci.v1beta1.ResponseInfo

* proto: restore serialization on Info response fields

The custom serialization attributes were not copied over from
v1beta1.ResponseInfo to v1.InfoResponse.

* proto: serde fallbacks for Vote extension fields

When decoding pre-0.38 messages, the "extension" and "extension_signature" fields are not present. Previously, this worked
because the v0_37 version of the generated struct was used; this,
on the other hand, resulted in the extension fields not being properly decoded in RPC with 0.38+ nodes.